Optimize failure notice of copyright header check (#23)
fix https://github.com/NVIDIA/spark-rapids-common/issues/20

If the feature branch is not based on the latest target branch, git diff
may include unexpected files marked as modified in the check that are
not from the feature branch. This is because it tries to find a common
commit between the two branches and can get confused if it is too old
and there are lots of changes. If this happens please upmerge your PR to
the latest development branch.

now the check is based on merged commit vs base ref, it should not detect untouched files
---------

Signed-off-by: Yanxuan Liu <yanxuanl@nvidia.com>